Add fade-in functionality to sign-up failure
Problem: When a user signs up, they receive the success message, which
fades. If they try to sign up again but have invalid credentials, the
failure message doesn't appear because the div was faded on success.

Solution:
When running 'signUpFailure', fade-in the div first. Then add the new
text.

Note to self: either remember this or use different divs for different
messages (success vs failure).