Fix fadeout issue for sign-out failure
Problem: if user signs in and signs out successfully, the message
appears as expected and fades. If the user signs in again and has a
sign out failure, the failure message will not appear because of the
success message fadeout.

Solution: fade-in div on signOutFailure so that a message will appear.

Again, use different message boxes for failure vs success. Or pay closer
attention.