Remove fade out functionality
Fade out functionality was causing more trouble than it was worth.
Decided to remove it entirely, because I couldn't get it to stop
midway through animation if it was interrupted with a message change to
display the new message at full opacity.

Also changed "Change Password" messages to not fade out. Same issue as
above - if you clicked the change password button multiple times, it
caused a chaining reaction of fading out and fading back in over and
over again. No bueno.

Cleared out the message on all clicks within the application. Messy,
sure, but it works.